==> notebooks/1. Data Preprocessing.ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np
from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/original/CSE_CIC_IDS2018/ddos_portmap_2018.csv")

ddos_ldap_2019_df = pd.read_csv("../data/original/CIC_DDoS2019/ddos_ldap_2019.csv")

ddos_netbios_2019_df = pd.read_csv("../data/original/CIC_DDoS2019/ddos_netbios_2019.csv")

ddos_syn_2019_df = pd.read_csv("../data/original/CIC_DDoS2019/ddos_syn_2019.csv")

ddos_udp_2019_df = pd.read_csv("../data/original/CIC_DDoS2019/ddos_udp_2019.csv")

target_feature = 'Label'

# 3. Removing Null Entries, Normalizing Columns, and Reducing the Size of the Datasets

The datasets are way too large to even store in GitHub. We will reduce the size while keeping the distribution of target feature.

## 3.1 PORTMAN_2018 Dataset

This dataset is not too large to save in GitHub. Only need to convert categorical features to numerical and save in new location. Also remove spaces from column titles, and remove time dependant features.

ddos_portmap_2018_df.info()

# Removing whitespace from column titles
new_ddos_portmap_2018_df = ddos_portmap_2018_df.copy()
new_ddos_portmap_2018_df.columns = ddos_portmap_2018_df.columns.str.replace(' ', '')

new_ddos_portmap_2018_df[target_feature].value_counts()

le = preprocessing.LabelEncoder()
le.fit(new_ddos_portmap_2018_df[target_feature])
new_ddos_portmap_2018_df[target_feature] = le.transform(new_ddos_portmap_2018_df[target_feature])

new_ddos_portmap_2018_df[target_feature].value_counts()

# Find features with empty cells

columns_with_nan = new_ddos_portmap_2018_df.columns[new_ddos_portmap_2018_df.isna().any()].tolist()
print("Columns with NaN values:", columns_with_nan)

count = np.isinf(new_ddos_portmap_2018_df).values.sum() 
print("Inf Values count: " + str(count) + " infinite values") 

# Removing empty and infinite cells

new_ddos_portmap_2018_df.replace([np.inf, -np.inf], np.nan, inplace=True)
new_ddos_portmap_2018_df = new_ddos_portmap_2018_df[new_ddos_portmap_2018_df['FlowBytes/s'].notna()]

columns_with_nan = new_ddos_portmap_2018_df.columns[new_ddos_portmap_2018_df.isna().any()].tolist()
print("Columns with NaN values:", columns_with_nan)

count = np.isinf(new_ddos_portmap_2018_df).values.sum() 
print("Inf Values count: " + str(count) + " infinite values") 

new_ddos_portmap_2018_df.info()

scaler = preprocessing.MinMaxScaler()
d = scaler.fit_transform(new_ddos_portmap_2018_df.values)
temp = new_ddos_portmap_2018_df[target_feature].tolist()

scaled_ddos_portmap_2018_df = pd.DataFrame(d, columns=new_ddos_portmap_2018_df.columns)
scaled_ddos_portmap_2018_df[target_feature] = temp

scaled_ddos_portmap_2018_df = scaled_ddos_portmap_2018_df.round(9)

temp = scaled_ddos_portmap_2018_df[target_feature].tolist()
for index in range(len(temp)):
    if np.isnan(temp[index]):
        #print("FOUND")
        print(index)

scaled_ddos_portmap_2018_df.min(axis=0) # will return min value of each column

scaled_ddos_portmap_2018_df.max(axis=0) # will return max value of each column

scaled_ddos_portmap_2018_df.to_csv("../data/processed/CSE_CIC_IDS2018/ddos_portmap_2018.csv")

scaled_ddos_portmap_2018_df.info()

train, test = train_test_split(scaled_ddos_portmap_2018_df, test_size=0.3)
train.to_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv")
test.to_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_test.csv")

## 3.2 LDAP_2019 Dataset

# PREPROCESSING PIPELINE:

def preprocess_ddos_2019 (ddos_df, remove_value, num_attacks):
    
    # Removing whitespace from column titles
    print("Removing whitespaces from column titles....")
    new_ddos_df = ddos_df.copy()
    new_ddos_df.columns = ddos_df.columns.str.replace(' ', '')
    
    # Drop unnamed columns
    print("Dropping unnamed columns....")
    new_ddos_df = new_ddos_df.loc[:, ~new_ddos_df.columns.str.contains('^Unnamed')]
    
    # Drop extra values
    print("Dropping extra values in target feature....")
    if remove_value!=None:
        new_ddos_df = new_ddos_df.drop(new_ddos_df[new_ddos_df[target_feature] == remove_value].index)
    
    # Drop time dependant variables
    print("Dropping time dependant variables....")
    new_ddos_df = new_ddos_df.drop('Timestamp', axis=1)
    
    # Categorical to numerical features
    print("Converting categorical features to numerical....")
    categorical_features = ['FlowID', 'SourceIP', 'DestinationIP', 'SimillarHTTP', target_feature]

    for categorical_feature in categorical_features:
        new_ddos_df[categorical_feature] = new_ddos_df[categorical_feature].astype("string")
        le = preprocessing.LabelEncoder()
        le.fit(new_ddos_df[categorical_feature])
        new_ddos_df[categorical_feature] = le.transform(new_ddos_df[categorical_feature])
        
    # Drop empty and inf values
    print("Dropping nan, inf, -inf values.....")
    new_ddos_df.replace([np.inf, -np.inf], np.nan, inplace=True)
    new_ddos_df.dropna(inplace=True)
    
    # Normalize
    print("Normalizing....")
    scaler = preprocessing.MinMaxScaler()
    d = scaler.fit_transform(new_ddos_df.values)
    temp = new_ddos_df[target_feature].tolist()


    scaled_ddos_df = pd.DataFrame(d, columns=new_ddos_df.columns)
    scaled_ddos_df[target_feature] = temp
        
    # Drop a lot of the attack instances
    print("Keep only a certain ammount of samples......")
    attack_indices = scaled_ddos_df[scaled_ddos_df[target_feature] == 1].index
    benign_indices = scaled_ddos_df[scaled_ddos_df[target_feature] == 0].index
    random_indices = np.random.choice(attack_indices, num_attacks, replace=False)
    all_indices = np.concatenate((benign_indices, random_indices))
    scaled_ddos_df = scaled_ddos_df.loc[all_indices]
    
    # Round values
    print("Rounding values......")
    scaled_ddos_df = scaled_ddos_df.round(9)
    
    # Shuffle entries
    print("Shuffling....")
    scaled_ddos_df = shuffle(scaled_ddos_df)
    
    print("DONE!....")
    return scaled_ddos_df

ddos_ldap_2019_df.info()

ddos_ldap_2019_df[' Label'].value_counts()

new_ddos_ldap_2019_df = preprocess_ddos_2019(ddos_ldap_2019_df, 'NetBIOS', 200000)

# Find features with empty cells

columns_with_nan = new_ddos_ldap_2019_df.columns[new_ddos_ldap_2019_df.isna().any()].tolist()
print("Columns with NaN values:", columns_with_nan)

count = np.isinf(new_ddos_ldap_2019_df).values.sum() 
print("Inf Values count: " + str(count) + " infinite values") 

new_ddos_ldap_2019_df.min(axis=0) # will return min value of each column

new_ddos_ldap_2019_df.max(axis=0) # will return max value of each column

new_ddos_ldap_2019_df.info()

new_ddos_ldap_2019_df['Label'].value_counts()

new_ddos_ldap_2019_df.to_csv("../data/processed/CIC_DDoS2019/ddos_ldap_2019.csv")

train, test = train_test_split(new_ddos_ldap_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_test.csv")

## 3.3 NETBIOS_2019 Dataset

ddos_netbios_2019_df[' Label'].value_counts()

new_ddos_netbios_2019_df = preprocess_ddos_2019(ddos_netbios_2019_df, None, 200000)

new_ddos_netbios_2019_df['Label'].value_counts()

new_ddos_netbios_2019_df.to_csv("../data/processed/CIC_DDoS2019/ddos_netbios_2019.csv")

train, test = train_test_split(new_ddos_netbios_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_test.csv")

## 3.4 UDP_2019 Dataset

ddos_udp_2019_df[' Label'].value_counts()

new_ddos_udp_2019_df = preprocess_ddos_2019(ddos_udp_2019_df, None, 3134)
new_ddos_udp_2019_df['Label'].value_counts()

new_ddos_udp_2019_df.to_csv("../data/processed/CIC_DDoS2019/ddos_udp_2019.csv")

train, test = train_test_split(new_ddos_udp_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_test.csv")

## 3.5 SYN_2019 Dataset

ddos_syn_2019_df[' Label'].value_counts()

new_ddos_syn_2019_df = preprocess_ddos_2019(ddos_syn_2019_df, None, 100000)
new_ddos_syn_2019_df['Label'].value_counts()

new_ddos_syn_2019_df.to_csv("../data/processed/CIC_DDoS2019/ddos_syn_2019.csv")

train, test = train_test_split(new_ddos_syn_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_test.csv")

# DONE
