==> notebooks/2. Train Test Split.ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np
from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/processed/CSE_CIC_IDS2018/ddos_portmap_2018.csv")

ddos_portmap_2018_df.head(3)

ddos_ldap_2019_df = pd.read_csv("../data/processed/CIC_DDoS2019/ddos_ldap_2019.csv")

ddos_ldap_2019_df.head(3)

ddos_netbios_2019_df = pd.read_csv("../data/processed/CIC_DDoS2019/ddos_netbios_2019.csv")

ddos_netbios_2019_df.head(3)

ddos_syn_2019_df = pd.read_csv("../data/processed/CIC_DDoS2019/ddos_syn_2019.csv")

ddos_syn_2019_df.head(3)

ddos_udp_2019_df = pd.read_csv("../data/processed/CIC_DDoS2019/ddos_udp_2019.csv")

ddos_udp_2019_df.head(3)

target_feature = 'Label'

# 3. Split into Training and Testing Set for Evaluation

train, test = train_test_split(ddos_portmap_2018_df, test_size=0.3)
train.to_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv")
test.to_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_test.csv")

train, test = train_test_split(ddos_ldap_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_test.csv")

train, test = train_test_split(ddos_netbios_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_test.csv")

train, test = train_test_split(ddos_syn_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_test.csv")

train, test = train_test_split(ddos_udp_2019_df, test_size=0.3)
train.to_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv")
test.to_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_test.csv")

# DONE
