==> notebooks/2.1 Feature Selection VarianceThreshold (Alexandra).ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np
from sklearn.feature_selection import VarianceThreshold

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv", index_col=0)

ddos_ldap_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv", index_col=0)

ddos_netbios_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv", index_col=0)

ddos_syn_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv", index_col=0)

ddos_udp_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv", index_col=0)

target_feature = 'Label'

# 3. Feature Selection - Variance Threshold

### Helper Functions

def VarianceThreshold_Helper(X_train, threshold):
    
    new_X_train = X_train.copy()

    constant_filter = VarianceThreshold(threshold= threshold)
    constant_filter.fit(new_X_train)

    len(new_X_train.columns[constant_filter.get_support()])

    constant_columns = [column for column in new_X_train.columns
                        if column not in new_X_train.columns[constant_filter.get_support()]]

    new_X_train.drop(labels=constant_columns, axis=1, inplace=True)
    
    return new_X_train.columns.tolist()

def stable_VarianceThreshold(train_df, num_splits, threshold, verbose):
    
    selectedFeatures = []
    
    df_split = np.array_split(train_df, num_splits)
    
    for small_df in df_split:
        X_train = small_df.drop([target_feature], axis=1)
        y_train = small_df[target_feature]
        
        selectedFeatures = selectedFeatures + VarianceThreshold_Helper(X_train, threshold)
        # NOTE TO ARYAN, PRANAV, AND ANISHA: THIS IS THE LINE YOU SHOULD CHANGE, AFTER IMPLEMENTING YOUR FEATURE
        # SELECTION METHOD
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature))+"/"+str(num_splits))
        ranks.append(selectedFeatures.count(feature)/num_splits)
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data) 
    
    return rank_df

### Investigating a Good Choice of Threshold

X_train_portmap = ddos_portmap_2018_df.drop([target_feature], axis=1)
y_train_portmap = ddos_portmap_2018_df[target_feature]

X_train_portmap.columns

for t in [0.00001, 0.0001, 0.001, 0.01, 0.02, 0.05, 0.1, 0.2, 0.3]:
    features = VarianceThreshold_Helper(X_train = X_train_portmap, threshold = t)
    print("Variance Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
    print("_______________________________________________________________________________")

# 4. Saving the Rankings

portmap_ranking = stable_VarianceThreshold(ddos_portmap_2018_df, 10, 0.00005, verbose=False)
portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_VarianceThreshold.csv")
portmap_ranking.head(20)

ldap_ranking = stable_VarianceThreshold(ddos_ldap_2019_df, 10, 0.00005, verbose=False)
ldap_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_ldap_2019_VarianceThreshold.csv")
ldap_ranking.head(20)

netbios_ranking = stable_VarianceThreshold(ddos_netbios_2019_df, 10, 0.00005, verbose=False)
netbios_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_netbios_2019_VarianceThreshold.csv")
netbios_ranking.head(20)

syn_ranking = stable_VarianceThreshold(ddos_syn_2019_df, 10, 0.00005, verbose=False)
syn_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_syn_2019_VarianceThreshold.csv")
syn_ranking.head(20)

udp_ranking = stable_VarianceThreshold(ddos_udp_2019_df, 10, 0.00005, verbose=False)
udp_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_udp_2019_VarianceThreshold.csv")
udp_ranking.head(20)

# 5. Baselines

def baseline_VarianceThreshold(train_df, threshold, verbose):
    
    selectedFeatures = []
    
    X_train = train_df.drop([target_feature], axis=1)
    y_train = train_df[target_feature]
        
    selectedFeatures =  VarianceThreshold_Helper(X_train, threshold)
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature)))
        ranks.append(selectedFeatures.count(feature))
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data)
    
    return rank_df

portmap_ranking_baseline = baseline_VarianceThreshold(ddos_portmap_2018_df, 0.005, verbose=False)
portmap_ranking_baseline.to_csv("../baseline_ranking/CSE_CIC_IDS2018/ddos_portmap_2018_VarianceThreshold.csv")
portmap_ranking_baseline.head(5)

ldap_ranking_baseline = baseline_VarianceThreshold(ddos_ldap_2019_df, 0.005, verbose=False)
ldap_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_ldap_2019_VarianceThreshold.csv")
ldap_ranking_baseline.head(5)

udp_ranking_baseline = baseline_VarianceThreshold(ddos_udp_2019_df, 0.005, verbose=False)
udp_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_udp_2019_VarianceThreshold.csv")
udp_ranking_baseline.head(5)

netbios_ranking_baseline = baseline_VarianceThreshold(ddos_netbios_2019_df, 0.005, verbose=False)
netbios_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_netbios_2019_VarianceThreshold.csv")
netbios_ranking_baseline.head(5)

syn_ranking_baseline = baseline_VarianceThreshold(ddos_syn_2019_df, 0.005, verbose=False)
syn_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_syn_2019_VarianceThreshold.csv")
syn_ranking_baseline.head(5)

# 6. Repeat for Assignment Datasets 

a2_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment2_train.csv", index_col=0)
a3_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment3_train.csv", index_col=0)

a2ranking = stable_VarianceThreshold(a2_train, 10, 0.0005, verbose=False)
a2ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment2_VarianceThreshold.csv")
a2ranking.head(60)

a3ranking = stable_VarianceThreshold(a3_train, 10, 0.1, verbose=False)
a3ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment3_VarianceThreshold.csv")
a3ranking.head(20)

# Baselines 

a2_baseline = baseline_VarianceThreshold(a2_train, 0.005, verbose=False)
a2_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment2_VarianceThreshold.csv")
a2_baseline.head(5)

a3_baseline = baseline_VarianceThreshold(a3_train, 0.1, verbose=False)
a3_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment3_VarianceThreshold.csv")
a3_baseline.head(5)
