==> notebooks/2.2 Feature Selection SelectKBest (Aryan).ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np
from sklearn.feature_selection import SelectKBest
from sklearn.feature_selection import chi2

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv", index_col=0)

ddos_ldap_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv", index_col=0)

ddos_netbios_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv", index_col=0)

ddos_syn_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv", index_col=0)

ddos_udp_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv", index_col=0)

target_feature = 'Label'

# 3. Feature Selection Method

def SelectKBest_Helper(X_train, y_train, k, features):
    
    select = SelectKBest(score_func=chi2, k=k)
    z = select.fit_transform(X_train,y_train)
    
    myFilter = select.get_support()

    result = [feature for feature, selected in zip(features, myFilter) if selected]
    
    return result

def stable_SelectKBest(train_df, num_splits, k, verbose):
    
    selectedFeatures = []
    
    if num_splits != None:
        df_split = np.array_split(train_df, num_splits)
    
    for small_df in df_split:
        X_train = small_df.drop([target_feature], axis=1)
        y_train = small_df[target_feature]
    
        features = small_df.columns.tolist()
        
        selectedFeatures = selectedFeatures + SelectKBest_Helper(X_train, y_train, k, features)
        # NOTE TO ARYAN, PRANAV, AND ANISHA: THIS IS THE LINE YOU SHOULD CHANGE, AFTER IMPLEMENTING YOUR FEATURE
        # SELECTION METHOD
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature))+"/"+str(num_splits))
        ranks.append(selectedFeatures.count(feature)/num_splits)
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data) 
    
    return rank_df

portmap_ranking = stable_SelectKBest(ddos_portmap_2018_df, 10, 40, verbose=False)
portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_SelectKBest.csv")
portmap_ranking.head(20)

ldap_ranking = stable_SelectKBest(ddos_ldap_2019_df, 10, 40, verbose=False)
ldap_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_ldap_2019_SelectKBest.csv")
ldap_ranking.head(20)

netbios_ranking = stable_SelectKBest(ddos_netbios_2019_df, 10, 40, verbose=False)
netbios_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_netbios_2019_SelectKBest.csv")
netbios_ranking.head(20)

syn_ranking = stable_SelectKBest(ddos_syn_2019_df, 10, 40, verbose=False)
syn_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_syn_2019_SelectKBest.csv")
syn_ranking.head(20)

udp_ranking = stable_SelectKBest(ddos_udp_2019_df, 10, 40, verbose=False)
udp_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_udp_2019_SelectKBest.csv")
udp_ranking.head(20)

# 4. Baselines

def baseline_SelectKBest(train_df, k, verbose):
    
    # Take a smaller piece of the entire dataset
    df_split = np.array_split(train_df, 20)
    train_df = df_split[0]
    
    selectedFeatures = []
    
    X_train = train_df.drop([target_feature], axis=1)
    y_train = train_df[target_feature]
        
    features = train_df.columns.tolist()
    selectedFeatures =  SelectKBest_Helper(X_train, y_train, k, features)
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature)))
        ranks.append(selectedFeatures.count(feature))
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data)
    
    return rank_df

portmap_ranking_baseline = baseline_SelectKBest(ddos_portmap_2018_df, 40, verbose=False)
portmap_ranking_baseline.to_csv("../baseline_ranking/CSE_CIC_IDS2018/ddos_portmap_2018_SelectKBest.csv")
portmap_ranking_baseline.head(5)

ldap_ranking_baseline = baseline_SelectKBest(ddos_ldap_2019_df, 40, verbose=False)
ldap_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_ldap_2019_SelectKBest.csv")
ldap_ranking_baseline.head(5)

udp_ranking_baseline = baseline_SelectKBest(ddos_udp_2019_df, 40, verbose=False)
udp_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_udp_2019_SelectKBest.csv")
udp_ranking_baseline.head(5)

netbios_ranking_baseline = baseline_SelectKBest(ddos_netbios_2019_df, 40, verbose=False)
netbios_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_netbios_2019_SelectKBest.csv")
netbios_ranking_baseline.head(5)

syn_ranking_baseline = baseline_SelectKBest(ddos_syn_2019_df, 40, verbose=False)
syn_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_syn_2019_SelectKBest.csv")
syn_ranking_baseline.head(5)

# 5. Repeat with Assignment Datasets

a2_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment2_train.csv", index_col=0)
a3_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment3_train.csv", index_col=0)

a2ranking = stable_SelectKBest(a2_train, 10, 25, verbose=False)
a2ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment2_SelectKBest.csv")
a2ranking.head(20)

a3ranking = stable_SelectKBest(a3_train, 10, 4, verbose=False)
a3ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment3_SelectKBest.csv")
a3ranking.head(10)

# Baselines 

a2_baseline = baseline_SelectKBest(a2_train, 25, verbose=False)
a2_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment2_SelectKBest.csv")
a2_baseline.head(5)

a3_baseline = baseline_SelectKBest(a3_train, 4, verbose=False)
a3_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment3_SelectKBest.csv")
a3_baseline.head(5)
