==> notebooks/2.3 Feature Selection MutualInformation (Pranav).ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv", index_col=0)

ddos_ldap_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv", index_col=0)

ddos_netbios_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv", index_col=0)

ddos_syn_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv", index_col=0)

ddos_udp_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv", index_col=0)

target_feature = 'Label'

# 3. Feature Selection Method

from sklearn.feature_selection import mutual_info_classif

def MutualInformation_Helper(X_train, y_train, threshold):
    new_X_train = X_train.copy()

    # Calculate mutual information between each feature and the target variable
    mi_scores = mutual_info_classif(new_X_train, y_train)

    # Create a boolean mask to filter features based on the threshold
    selected_features_mask = mi_scores > threshold

    # Select features based on the mask
    new_X_train = new_X_train.loc[:, selected_features_mask]

    return new_X_train.columns.tolist()

def stable_MutualInformation(train_df, num_splits, threshold, verbose):
    
    selectedFeatures = []
    
    df_split = np.array_split(train_df, num_splits)
    
    for small_df in df_split:
        X_train = small_df.drop([target_feature], axis=1)
        y_train = small_df[target_feature]
        
        selectedFeatures = selectedFeatures + MutualInformation_Helper(X_train, y_train, threshold)
       
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature))+"/"+str(num_splits))
        ranks.append(selectedFeatures.count(feature)/num_splits)
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data) 
    
    return rank_df

portmap_ranking = stable_MutualInformation(ddos_syn_2019_df, 10, 0.004, verbose=False)

portmap_ranking.head(20)

portmap_ranking = stable_MutualInformation(ddos_portmap_2018_df, 10, 0.1, verbose=False)
portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_MutualInformation.csv")
portmap_ranking.head(20)

# 4. Saving the Ranking

CODE TO SAVE A CSV FILE WITH:
- 'feature' is a column of all of the features in the original dataset
- 'ranking' is a column with the rankings (in range 0 to 1) of the feature

portmap_ranking = stable_MutualInformation(ddos_portmap_2018_df, 10, 0.004, verbose=False)
portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_MutualInformation.csv")
portmap_ranking.head(20)

ldap_ranking = stable_MutualInformation(ddos_ldap_2019_df, 10, 0.05, verbose=False)
ldap_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_ldap_2019_MutualInformation.csv")
ldap_ranking.head(20)

netbios_ranking = stable_MutualInformation(ddos_netbios_2019_df, 10, 0.009, verbose=False)
netbios_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_netbios_2019_MutualInformation.csv")
netbios_ranking.head(20)

syn_ranking = stable_MutualInformation(ddos_syn_2019_df, 10, 0.008, verbose=False)
syn_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_syn_2019_MutualInformation.csv")
syn_ranking.head(20)

udp_ranking = stable_MutualInformation(ddos_udp_2019_df, 10, 0.08, verbose=False)
udp_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_udp_2019_MutualInformation.csv")
udp_ranking.head(20)

# 5. Baselines

def baseline_MutualInformation(train_df, threshold, verbose):
    
    selectedFeatures = []
    
    X_train = train_df.drop([target_feature], axis=1)
    y_train = train_df[target_feature]
        
    selectedFeatures =  MutualInformation_Helper(X_train, y_train, threshold)
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature)))
        ranks.append(selectedFeatures.count(feature))
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data)
    
    return rank_df

portmap_ranking_baseline = baseline_MutualInformation(ddos_portmap_2018_df, 0.004, verbose=False)
portmap_ranking_baseline.to_csv("../baseline_ranking/CSE_CIC_IDS2018/ddos_portmap_2018_MutualInformation.csv")
portmap_ranking_baseline.head(5)

ldap_ranking_baseline = baseline_MutualInformation(ddos_ldap_2019_df, 0.05, verbose=False)
ldap_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_ldap_2019_MutualInformation.csv")
ldap_ranking_baseline.head(5)

udp_ranking_baseline = baseline_MutualInformation(ddos_udp_2019_df, 0.08, verbose=False)
udp_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_udp_2019_MutualInformation.csv")
udp_ranking_baseline.head(5)

netbios_ranking_baseline = baseline_MutualInformation(ddos_netbios_2019_df, 0.009, verbose=False)
netbios_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_netbios_2019_MutualInformation.csv")
netbios_ranking_baseline.head(5)

syn_ranking_baseline = baseline_MutualInformation(ddos_syn_2019_df, 0.008, verbose=False)
syn_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_syn_2019_MutualInformation.csv")
syn_ranking_baseline.head(5)

# 6. Repeating with Assignment Datasets

a2_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment2_train.csv", index_col=0)
a3_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment3_train.csv", index_col=0)

a2ranking = stable_MutualInformation(a2_train, 10, 0.008, verbose=False)
a2ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment2_MutualInformation.csv")
a2ranking.head(20)

a3ranking = stable_MutualInformation(a3_train, 10, 0.08, verbose=False)
a3ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment3_MutualInformation.csv")
a3ranking.head(10)

# Baselines 

a2_baseline = baseline_MutualInformation(a2_train, 0.008, verbose=False)
a2_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment2_MutualInformation.csv")
a2_baseline.head(5)

a3_baseline = baseline_MutualInformation(a3_train, 0.08, verbose=False)
a3_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment3_MutualInformation.csv")
a3_baseline.head(5)
