==> notebooks/2.5 Feature Selection VarianceInflationFactor (Alexandra).ipynb <==
# 1. Imports

import pandas as pd
from sklearn import preprocessing
import numpy as np
from sklearn.linear_model import LinearRegression

# 2. Importing the Datasets

ddos_portmap_2018_df = pd.read_csv("../data/train_test/CSE_CIC_IDS2018/ddos_portmap_2018_train.csv", index_col=0)

ddos_ldap_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_ldap_2019_train.csv", index_col=0)

ddos_netbios_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_netbios_2019_train.csv", index_col=0)

ddos_syn_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_syn_2019_train.csv", index_col=0)

ddos_udp_2019_df = pd.read_csv("../data/train_test/CIC_DDoS2019/ddos_udp_2019_train.csv", index_col=0)

target_feature = 'Label'

# 3. Feature Selection - Variance Inflation Factor

### Helper Functions

def VIF_Helper(X_train, threshold):
    
    new_X_train = X_train.copy()
    features = new_X_train.columns.tolist()
    
    selected_features = []

    for feature in features:
        X = [f for f in features if f != feature]
        X, y = X_train[X], X_train[feature]
        r2 = LinearRegression().fit(X, y).score(X, y)
        if r2==1.0:
            vif = 50
        else:
            vif = 1/(1 - r2)
            
        #print(feature)
        #print(vif)
            
        if vif < threshold:
            selected_features.append(feature)

    return selected_features

def stable_VIF(train_df, num_splits, threshold, verbose):
    
    selectedFeatures = []
    
    df_split = np.array_split(train_df, num_splits)
    
    for small_df in df_split:
        X_train = small_df.drop([target_feature], axis=1)
        y_train = small_df[target_feature]
        
        selectedFeatures = selectedFeatures + VIF_Helper(X_train, threshold)
        # NOTE TO ARYAN, PRANAV, AND ANISHA: THIS IS THE LINE YOU SHOULD CHANGE, AFTER IMPLEMENTING YOUR FEATURE
        # SELECTION METHOD
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature))+"/"+str(num_splits))
        ranks.append(selectedFeatures.count(feature)/num_splits)
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data) 
    
    return rank_df

### Investigating a Good Choice of Threshold

X_train_portmap = ddos_portmap_2018_df.drop([target_feature], axis=1)
y_train_portmap = ddos_portmap_2018_df[target_feature]

len(X_train_portmap.columns.tolist())

- 1 — features are not correlated
- 1<VIF<5 — features are moderately correlated
- VIF>5 — features are highly correlated
- VIF>10 — high correlation between features and is cause for concern

t=10.0
features = VIF_Helper(X_train_portmap, threshold = t)
print("NonCollinear Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
print("_______________________________________________________________________________")

t=9.0
features = VIF_Helper(X_train_portmap, threshold = t)
print("NonCollinear Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
print("_______________________________________________________________________________")

t=5.0
features = VIF_Helper(X_train_portmap, threshold = t)
print("NonCollinear Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
print("_______________________________________________________________________________")

t=4.0
features = VIF_Helper(X_train_portmap, threshold = t)
print("NonCollinear Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
print("_______________________________________________________________________________")

t=1.0
features = VIF_Helper(X_train_portmap, threshold = t)
print("NonCollinear Threshold "+str(t)+": "+". Features: ("+str(len(features))+") "+str(features))
print("_______________________________________________________________________________")

# 4. Saving the Rankings

portmap_ranking = stable_VIF(ddos_portmap_2018_df, 10, 10.0, verbose=False)
portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_VIF.csv")
portmap_ranking.head(20)

portmap_ranking.to_csv("../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_VIF.csv")

ldap_ranking = stable_VIF(ddos_ldap_2019_df, 10, 10.0, verbose=False)
ldap_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_ldap_2019_VIF.csv")
ldap_ranking.head(20)

netbios_ranking = stable_VIF(ddos_netbios_2019_df, 10, 10.0, verbose=False)
netbios_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_netbios_2019_VIF.csv")
netbios_ranking.head(20)

syn_ranking = stable_VIF(ddos_syn_2019_df, 10, 10.0, verbose=False)
syn_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_syn_2019_VIF.csv")
syn_ranking.head(20)

udp_ranking = stable_VIF(ddos_udp_2019_df, 10, 10.0, verbose=False)
udp_ranking.to_csv("../ranking/CIC_DDoS2019/ddos_udp_2019_VIF.csv")
udp_ranking.head(20)

# 5. Baselines

def baseline_VIF(train_df, threshold, verbose):
    
    selectedFeatures = []
    
    X_train = train_df.drop([target_feature], axis=1)
    y_train = train_df[target_feature]
        
    selectedFeatures =  VIF_Helper(X_train, threshold)
        
    rank_data = {}
    features = train_df.columns.tolist()
    ranks = []
        
    for feature in features:
        if verbose :
            print("Feature: "+feature+". Count: "+str(selectedFeatures.count(feature)))
        ranks.append(selectedFeatures.count(feature))
    
    rank_data = {'feature':features, 'rank':ranks}
    rank_df = pd.DataFrame(rank_data)
    
    return rank_df

portmap_ranking_baseline = baseline_VIF(ddos_portmap_2018_df, 20.0, verbose=False)
portmap_ranking_baseline.to_csv("../baseline_ranking/CSE_CIC_IDS2018/ddos_portmap_2018_VIF.csv")
portmap_ranking_baseline.head(5)

ldap_ranking_baseline = baseline_VIF(ddos_ldap_2019_df, 20.0, verbose=False)
ldap_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_ldap_2019_VIF.csv")
ldap_ranking_baseline.head(5)

udp_ranking_baseline = baseline_VIF(ddos_udp_2019_df, 20.0, verbose=False)
udp_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_udp_2019_VIF.csv")
udp_ranking_baseline.head(5)

netbios_ranking_baseline = baseline_VIF(ddos_netbios_2019_df, 20.0, verbose=False)
netbios_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_netbios_2019_VIF.csv")
netbios_ranking_baseline.head(5)

syn_ranking_baseline = baseline_VIF(ddos_syn_2019_df, 20.0, verbose=False)
syn_ranking_baseline.to_csv("../baseline_ranking/CIC_DDoS2019/ddos_syn_2019_VIF.csv")
syn_ranking_baseline.head(5)

# 6. Repeating with Assignment Datasets

a2_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment2_train.csv", index_col=0)
a3_train = pd.read_csv("../data/csi5388_assignment2_3_data/assignment3_train.csv", index_col=0)

a2ranking = stable_VIF(a2_train, 10, 20.0, verbose=False)
a2ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment2_VIF.csv")
a2ranking.head(20)

a3ranking = stable_VIF(a3_train, 10, 39.0, verbose=False)
a3ranking.to_csv("../ranking/csi5388_assignment2_3_data/assignment3_VIF.csv")
a3ranking.head(10)

# Baselines 

a2_baseline = baseline_VIF(a2_train, 20, verbose=False)
a2_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment2_VIF.csv")
a2_baseline.head(5)

a3_baseline = baseline_VIF(a3_train, 39.0, verbose=False)
a3_baseline.to_csv("../baseline_ranking/csi5388_assignment2_3_data/assignment3_VIF.csv")
a3_baseline.head(5)
