==> notebooks/3. STEF-Rank.ipynb <==
# 1. Imports

import pandas as pd
import numpy as np
import glob
import os

# 2. Importing the Ranks

path_to_ranking = r'../ranking/CSE_CIC_IDS2018/ddos_portmap_2018_*.csv'

dataframes_portmap = []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_portmap.append(df)

path_to_ranking = r'../ranking/CIC_DDoS2019/ddos_ldap_2019_*.csv'

dataframes_ldap= []
#features = []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    feature = df['feature'].tolist()
    #print(df['feature'].tolist())
    #print(len(df['feature'].tolist()))
    #print(df['feature'].isna().sum())
    dataframes_ldap.append(df)
    #features.append(feature)
    
#all_same = all(len(lst) == len(features[0]) and all(a == b for a, b in zip(lst, features[0])) for lst in features)

#print("All lists are the same:", all_same)

path_to_ranking = r'../ranking/CIC_DDoS2019/ddos_udp_2019_*.csv'

dataframes_udp= []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_udp.append(df)

path_to_ranking = r'../ranking/CIC_DDoS2019/ddos_netbios_2019_*.csv'

dataframes_netbios= []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_netbios.append(df)

path_to_ranking = r'../ranking/CIC_DDoS2019/ddos_syn_2019_*.csv'

dataframes_syn= []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_syn.append(df)

# 3. Creating STEF-Rank Results

def STEF_Rank(ranking_dfs):
    
    aggregated_data = []
    
    features = ranking_dfs[0]['feature']
    
    for feature in features:
        
        total_ranking = 0
        
        # Add up the rankings
        for method in ranking_dfs:
            
            total_ranking += method.loc[method['feature'] == feature, 'rank'].sum()
            
        # Save the final ranking    
        aggregated_data.append({'feature': feature, 'rank': total_ranking/len(ranking_dfs)})
        
    # Create Dataframe    
    result_df = pd.DataFrame(columns=['feature', 'rank'])
    result_df['feature'] = pd.Series(dtype='str')
    result_df['rank'] = pd.Series(dtype='float')
    
    result_df = pd.DataFrame(aggregated_data)
        
    return result_df

stef_rank_portmap = STEF_Rank(dataframes_portmap)
stef_rank_portmap.to_csv("../stef_ranking/CSE_CIC_IDS2018/ddos_portmap_2018.csv")
stef_rank_portmap.head(3)

stef_rank_ldap = STEF_Rank(dataframes_ldap)
stef_rank_ldap.to_csv("../stef_ranking/CIC_DDoS2019/ddos_ldap_2019.csv")
stef_rank_ldap.head(3)

stef_rank_udp = STEF_Rank(dataframes_udp)
stef_rank_udp.to_csv("../stef_ranking/CIC_DDoS2019/ddos_udp_2019.csv")
stef_rank_udp.head(3)

stef_rank_netbios = STEF_Rank(dataframes_netbios)
stef_rank_netbios.to_csv("../stef_ranking/CIC_DDoS2019/ddos_netbios_2019.csv")
stef_rank_netbios.head(3)

stef_rank_syn = STEF_Rank(dataframes_syn)
stef_rank_syn.to_csv("../stef_ranking/CIC_DDoS2019/ddos_syn_2019.csv")
stef_rank_syn.head(3)

# 4. Repeating with Assignment Datasets

path_to_ranking = r'../ranking/csi5388_assignment2_3_data/assignment2_*.csv'

dataframes_a2 = []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_a2.append(df)

path_to_ranking = r'../ranking/csi5388_assignment2_3_data/assignment3_*.csv'

dataframes_a3 = []

for file in glob.glob(path_to_ranking):
    print(f"Processing file: {file}")
    df = pd.read_csv(file, index_col=0)
    df = df[~df['feature'].str.contains('Unnamed', na=False)]
    dataframes_a3.append(df)

stef_rank_syn = STEF_Rank(dataframes_a2)
stef_rank_syn.to_csv("../stef_ranking/csi5388_assignment2_3_data/assignment2.csv")
stef_rank_syn.head(3)

stef_rank_syn = STEF_Rank(dataframes_a3)
stef_rank_syn.to_csv("../stef_ranking/csi5388_assignment2_3_data/assignment3.csv")
stef_rank_syn.head(3)
